fix: fix jpush
