+++ packages/kit/src/provider/Container/NotificationHandlerContainer/hooks.native.ts
@@ -4,11 +4,15 @@ import type { RefObject } from 'react';
 import launchOptionsManager from '@onekeyhq/shared/src/modules/LaunchOptionsManager';
 import type { INavigateToNotificationDetailParams } from '@onekeyhq/shared/src/utils/notificationsUtils';
 import notificationsUtils from '@onekeyhq/shared/src/utils/notificationsUtils';
-import type { INotificationPushMessageInfo } from '@onekeyhq/shared/types/notification';
+import type {
+  IJPushRemotePushMessageInfo,
+  INotificationPushMessageInfo,
+} from '@onekeyhq/shared/types/notification';
 
 import backgroundApiProxy from '../../../background/instance/backgroundApiProxy';
 import { useVersionCompatible } from '../../../hooks/useVersionCompatible';
 import { whenAppUnlocked } from '../../../utils/passwordUtils';
+import { ColdStartByNotification } from '../ColdStartByNotification';
 
 import type { IAccountSelectorActiveAccountInfo } from '../../../states/jotai/contexts/accountSelector';
 
@@ -50,6 +54,49 @@ export const useInitialNotification = (
     setTimeout(async () => {
       if (coldStartRef.current) {
         coldStartRef.current = false;
+        const options: IJPushRemotePushMessageInfo | null =
+          ColdStartByNotification.launchNotification;
+        if (options) {
+          console.log(
+            'coldStart ColdStartByNotification launchNotification',
+            options,
+          );
+          options.msgId =
+            options?.params?.msgId ||
+            options?.msgId ||
+            options?._j_msgid?.toString() ||
+            '';
+          console.log(
+            'coldStart ColdStartByNotification launchNotification FIXED',
+            options,
+          );
+          const title = options.aps?.alert?.title || '';
+          const content = options.aps?.alert?.body || '';
+          const icon = options?.image;
+          const badge = options.aps?.badge?.toString() || '';
+
+          void backgroundApiProxy.serviceNotification.handleColdStartByNotification(
+            {
+              notificationId: options.msgId,
+              params: {
+                notificationId: options.msgId,
+                title,
+                description: content,
+                icon,
+                remotePushMessageInfo: {
+                  pushSource: 'jpush',
+                  title,
+                  content,
+                  badge,
+                  extras: {
+                    ...options,
+                  },
+                },
+              },
+            },
+          );
+          return;
+        }
         const launchOptions = await launchOptionsManager.getLaunchOptions();
         let userInfo: INotificationPushMessageInfo | undefined;
         if (launchOptions?.localNotification) {
